refactor: ♻️ Refactor `UserRole` and `UserSnapshot` typeDefs
