+++ packages/common/src/types/userSnapshot/index.js
@@ -1,12 +1,16 @@
 import { make } from './creation';
-import { apiExample, dbExample } from './examples';
+import { apiExample, dbExample, uiExample } from './examples';
 import { dbToApi } from './transforms';
+import { propTypes, typeDef } from './typedef';
 
 const UserSnapshot = {
   apiExample,
   dbExample,
   dbToApi,
   make,
+  propTypes,
+  typeDef,
+  uiExample,
 };
 
 export { UserSnapshot };
